refactor: moved logging to utils
